+++ src/components/cards/textCard.js
